toggle and mock response data on dev env

+++ services/removeToFavs.ts
@@ -1,10 +1,18 @@
+import { USE_MOCKS } from '@/toggles'
+
 export const removeToFavs = async ({ token, id }: { token: string; id: string }) => {
-  const favsResponse = await fetch(`${process.env.nextApiURL}/films/user/list/${id}`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${token}`,
-      'Content-Type': 'application/json'
+  if (USE_MOCKS) {
+    return {
+      myList: []
     }
-  })
-  return await favsResponse.json()
+  } else {
+    const favsResponse = await fetch(`${process.env.nextApiURL}/films/user/list/${id}`, {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json'
+      }
+    })
+    return await favsResponse.json()
+  }
 }
